requirements.txt
absl-py==0.8.0
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
gast==0.2.2
gitdb2==2.0.5
GitPython==3.0.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.1
Pillow==6.1.0
progressbar2==3.43.1
protobuf==3.9.1
pyparsing==2.4.2
python-utils==2.3.0
PyYAML==5.1.2
requests==2.22.0
scipy==1.3.1
six==1.12.0
smmap2==2.0.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.5
wget==3.2
wrapt==1.11.2
